♻️ refactor: refactor the trpc client with react query mode
