+++ src/libs/trpc/client/index.ts
@@ -1,0 +1,2 @@
+export { edgeClient } from './edge';
+export * from './lambda';
